LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/synced-state.https-expected.txt


FAIL worker objects for the same entity have the same state assert_equals: getRegistration should return the same object expected object "[object ServiceWorkerRegistration]" but got null
